fix: change setting.request_max_price type as int

+++ sql/V4__create_settings_table.sql
@@ -7,7 +7,7 @@ CREATE TABLE `settings` (
           `loan_days` TINYINT unsigned NOT NULL,
           `loan_max_book` TINYINT unsigned NOT NULL,
           `request_max_count` TINYINT unsigned NOT NULL,
-          `request_max_price` TINYINT unsigned NOT NULL,
+          `request_max_price` INT unsigned NOT NULL,
           `created_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
           `updated_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
           `is_deleted` BOOLEAN NOT NULL DEFAULT FALSE
